backend/requirements.txt
Flask>=3.0
flask-cors>=4.0
pymongo>=4.8
pydantic>=2.8
python-dotenv>=1.0
